+++ src/blocksimpy/__init__.py
@@ -1,3 +1,3 @@
 """Blockchain Discrete Event Simulator."""
 
-__version__ = "1.3.0"
+__version__ = "1.4.0"